backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
supabase>=2.3.0
PyJWT>=2.8.0
python-multipart>=0.0.6
